+++ scalapb-runtime/src/main/scalajvm/com/google/protobuf/struct/Value.scala
@@ -216,7 +216,7 @@ object Value extends scalapb.GeneratedMessageCompanion[com.google.protobuf.struc
   lazy val defaultInstance = com.google.protobuf.struct.Value(
     kind = com.google.protobuf.struct.Value.Kind.Empty
   )
-  sealed trait Kind extends _root_.scalapb.GeneratedOneof {
+  sealed abstract class Kind extends _root_.scalapb.GeneratedOneof {
     def isEmpty: _root_.scala.Boolean = false
     def isDefined: _root_.scala.Boolean = true
     def isNullValue: _root_.scala.Boolean = false