+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Bump protobuf and protoc to 4.26.0. Currently files must specify proto2 and
   proto3 syntax. Editions are not supported yet.
 - References to descriptors are fully qualified (#1724)
+- Oneofs are now generated as sealed abstract class instead of sealed traits (#1694)
 
 ## [0.11.18] (Unreleased)
 - References to descriptors are fully qualified (#1724)